requirements.txt
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=9.0.0
gspread
